# public/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "HDFC SmartGateway API",
        "description": "API documentation for HDFC SmartGateway TranPortal integration.\n\n## Tags\n- **Standard Payment** — Create orders and check payment status\n- **Mandates** — Full mandate lifecycle: create, register, execute, list, status, revoke\n- **One-Time Mandate (OTM)** — Fund blocking for one-time debits\n- **CVV-less Payments** — Tokenized card checkout without CVV\n\n## Authentication\nAll calls use Basic HTTP Auth:\n```\nAuthorization: Basic <Base64(API_KEY:)>\n```\n\n## Environments\n| Env | Base URL |\n|---|---|\n| Sandbox | `https://smartgateway.hdfcuat.bank.in` |\n| Production | `https://smartgateway.hdfc.bank.in` |",
        "version": "2.0.0",
        "contact": {
            "name": "Exelon Development Team"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local Dev Server"
        }
    ],
...
    "paths": {
...
        "/api/surcharge/create-order": {
            "post": {
                "tags": [
                    "Surcharge (Convenience Fee)"
                ],
                "summary": "Create Order with Surcharge",
                "description": "Create an order with a convenience fee (surcharge) added on top of the order amount.\n\n**Default surcharge: ₹50 flat fee + ₹9 GST.**\n\n⚠️ Surcharge rules MUST be configured in the Juspay Dashboard first:\n`PG Control Center > Surcharge (4th tab) > Add New Rule`",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "description": "Base order amount (before surcharge)",
                                        "example": 500
                                    },
                                    "currency": {
                                        "type": "string",
                                        "default": "INR"
                                    },
                                    "customer_id": {
                                        "type": "string",
                                        "example": "cust_001"
                                    },
                                    "customer_email": {
                                        "type": "string",
                                        "example": "test@example.com"
                                    },
                                    "customer_phone": {
                                        "type": "string",
                                        "example": "9876543210"
                                    },
                                    "surcharge_amount": {
                                        "type": "string",
                                        "description": "Convenience fee amount (default: ₹50)",
                                        "default": "50.00",
                                        "example": "50.00"
                                    },
                                    "tax_amount": {
                                        "type": "string",
                                        "description": "Tax on surcharge (default: ₹9 = 18% GST)",
                                        "default": "9.00",
                                        "example": "9.00"
                                    },
                                    "return_url": {
                                        "type": "string"
                                    }
                                }
                            },
                            "examples": {
                                "with_default_surcharge": {
                                    "summary": "₹500 order + ₹50 surcharge + ₹9 tax = ₹559 total",
                                    "value": {
                                        "amount": 500,
                                        "customer_id": "cust_001",
                                        "customer_email": "test@example.com",
                                        "customer_phone": "9876543210"
                                    }
                                },
                                "with_custom_surcharge": {
                                    "summary": "₹1000 order + ₹100 surcharge + ₹18 tax = ₹1118 total",
                                    "value": {
                                        "amount": 1000,
                                        "surcharge_amount": "100.00",
                                        "tax_amount": "18.00",
                                        "customer_id": "cust_001",
                                        "customer_email": "test@example.com",
                                        "customer_phone": "9876543210"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created with surcharge breakdown"
                    }
                }
            }
        },
        "/api/surcharge/order/{orderId}": {
            "get": {
                "tags": [
                    "Surcharge (Convenience Fee)"
                ],
                "summary": "Get Surcharge Order Details",
                "description": "Retrieve order details with surcharge breakdown — shows base amount, surcharge, tax, and net amount.",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "surcharge_17720200703"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details with surcharge info"
                    }
                }
            }
        },
        "/api/surcharge/refund": {
            "post": {
                "tags": [
                    "Surcharge (Convenience Fee)"
                ],
                "summary": "Refund Surcharge Order",
                "description": "Refund an order that had surcharge applied.\n\nWhether surcharge is included in the refund depends on the Dashboard setting:\n- **Enabled**: Full refund includes surcharge\n- **Disabled**: Partial refund excluding surcharge",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "order_id"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "string",
                                        "description": "Order ID of the surcharge order to refund",
                                        "example": "surcharge_17720200703"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "description": "Custom refund amount (optional — auto-calculated if not provided)"
                                    },
                                    "include_surcharge": {
                                        "type": "boolean",
                                        "description": "Include surcharge in refund amount (default: false)",
                                        "default": false
                                    }
                                }
                            },
                            "examples": {
                                "refund_without_surcharge": {
                                    "summary": "Refund base amount only (exclude surcharge)",
                                    "value": {
                                        "order_id": "surcharge_17720200703"
                                    }
                                },
                                "refund_with_surcharge": {
                                    "summary": "Full refund including surcharge",
                                    "value": {
                                        "order_id": "surcharge_17720200703",
                                        "include_surcharge": true
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Refund initiated with breakdown"
                    }
                }
            }
        },
...
    },
...
}